[UPDATE] centralize stopPoll for button + command
